expanded polyhedral tests
