+++ tests/feature/hist/symmetries/cyclic_symmetry.cpp
@@ -6,7 +6,9 @@
 #include <data/Molecule.h>
 #include <data/symmetry/CyclicSymmetry.h>
 #include <data/symmetry/PointSymmetry.h>
-#include <data/symmetry/PolyhedralSymmetry.h>
+#include <data/symmetry/TetrahedralSymmetry.h>
+#include <data/symmetry/OctahedralSymmetry.h>
+#include <data/symmetry/IcosahedralSymmetry.h>
 #include <data/symmetry/CompositeSymmetry.h>
 #include <data/symmetry/ReferenceSymmetry.h>
 #include <hist/intensity_calculator/ICompositeDistanceHistogramExv.h>
@@ -28,47 +30,6 @@ auto make_unique_cyclic_sym = [] (
 ) {
     return std::make_unique<symmetry::CyclicSymmetry>(initial_relation, per_step_translation, axis, angle, repetitions);
 };
-
-struct RES {
-    RES(double d, int v) : index(std::round(d*settings::flags::inv_bin_width)), val(v) {}
-    int index;
-    int val;
-};
-
-void check_hist(const std::vector<double>& h, std::vector<RES> checks) {
-    std::sort(checks.begin(), checks.end(), [](const RES& a, const RES& b) {return a.index < b.index;});
-    std::vector<RES> tmp;
-    for (int i = 0; i < static_cast<int>(checks.size()); ++i) {
-        if (i == 0 || checks[i].index != checks[i-1].index) {
-            tmp.push_back(checks[i]);
-        } else {
-            tmp.back().val += checks[i].val;
-        }
-    }
-    checks = tmp;
-    REQUIRE(checks.back().index < static_cast<int>(h.size()));
-    int j = 0;
-    for (int i = 0; i < static_cast<int>(h.size()); ++i) {
-        if (i == checks[j].index) {
-            if (h[i] != checks[j].val) {
-                INFO("i = " << i << ", dist = " << i*constants::axes::d_axis.width());
-                INFO("h[i] = " << h[i] << ", checks[j].val = " << checks[j].val);
-                CHECK(false);
-            }
-            ++j;
-            if (j == static_cast<int>(checks.size())) {
-                break;
-            }
-        } else {
-            if (h[i] != 0) {
-                INFO("i = " << i << ", dist = " << i*constants::axes::d_axis.width());
-                INFO("h[i] = " << h[i] << ", checks[j].val = " << 0);
-                CHECK(false);
-            }
-        }
-    }
-    SUCCEED();
-}
 
 auto test_translation = [] (settings::hist::HistogramManagerChoice choice) {
     SECTION("one body with one atom") {
@@ -626,315 +587,5 @@ TEST_CASE("SymmetryManager: random tests") {
     }
     SECTION("PartialSymmetryManager") {
         test_random(settings::hist::HistogramManagerChoice::PartialHistogramSymmetryManagerMT);
-    }
-}
-
-auto make_unique_point_sym = [] (const Vector3<double>& translation, const Vector3<double>& rotation) {
-    return std::make_unique<symmetry::PointSymmetry>(translation, rotation);
-};
-
-auto test_point_symmetry = [] (settings::hist::HistogramManagerChoice choice) {
-    SECTION("one body with one atom") {
-        // Atom at its own cm so copy is always at atom_pos + translation, independent of rotation
-        AtomFF a({0, 0, 0}, form_factor::form_factor_t::C);
-        Molecule m({Body{std::vector{a}}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        SECTION("single copy via translation") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({1, 0, 0}, {0, 0, 0}));
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {RES(0, 2), RES(1, 2)});
-        }
-
-        SECTION("two copies via translation") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({1, 0, 0}, {0, 0, 0}));
-            m.get_body(0).symmetry().add(make_unique_point_sym({2, 0, 0}, {0, 0, 0}));
-            // atoms: {0,0,0}, {1,0,0}, {2,0,0}
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {RES(0, 3), RES(1, 4), RES(2, 2)});
-        }
-
-        SECTION("three copies via translation") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({1, 0, 0}, {0, 0, 0}));
-            m.get_body(0).symmetry().add(make_unique_point_sym({2, 0, 0}, {0, 0, 0}));
-            m.get_body(0).symmetry().add(make_unique_point_sym({3, 0, 0}, {0, 0, 0}));
-            // atoms: {0,0,0}, {1,0,0}, {2,0,0}, {3,0,0}
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {RES(0, 4), RES(1, 6), RES(2, 4), RES(3, 2)});
-        }
-
-        SECTION("rotation only (atom at cm: copy overlaps original)") {
-            // rotation={0,0,1} → Rz(1 rad); atom at cm={0,0,0}, no translation → copy lands on original
-            // 2 atoms at same position: 2 self-counts + 2 cross-pair counts = 4 at bin 0
-            m.get_body(0).symmetry().add(make_unique_point_sym({0, 0, 0}, {0, 0, 1}));
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {RES(0, 4)});
-        }
-
-        SECTION("combined translation and rotation (explicit_structure comparison)") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({2, 1, 0}, {0, 1, 0}));
-            auto h = m.get_histogram()->get_weighted_counts();
-
-            auto b = m.get_body(0).symmetry().explicit_structure();
-            auto m2 = Molecule({Body{std::move(b.atoms), std::move(b.waters)}});
-            set_unity_charge(m2);
-            auto h2 = m2.get_histogram()->get_weighted_counts();
-            CHECK(compare_hist_approx(h, h2));
-        }
-    }
-
-    SECTION("two bodies with one atom each") {
-        AtomFF a1({0, 0, 0}, form_factor::form_factor_t::C);
-        AtomFF a2({1, 0, 0}, form_factor::form_factor_t::C);
-        Molecule m({Body{std::vector{a1}}, Body{std::vector{a2}}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        SECTION("single copy of body1") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({0, 1, 0}, {0, 0, 0}));
-            // atoms: {0,0,0}, {1,0,0}, copy: {0,1,0}
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {RES(0, 3), RES(1, 4), RES(std::sqrt(2.0), 2)});
-        }
-
-        SECTION("copies on both bodies") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({0, 1, 0}, {0, 0, 0}));
-            m.get_body(1).symmetry().add(make_unique_point_sym({0, 1, 0}, {0, 0, 0}));
-            // atoms: {0,0,0}, {1,0,0}, copies: {0,1,0}, {1,1,0}
-            auto h = m.get_histogram()->get_weighted_counts();
-            check_hist(h, {
-                RES(0, 4),
-                RES(1, 8),
-                RES(std::sqrt(2.0), 4)
-            });
-        }
-
-        SECTION("rotation with explicit_structure comparison") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({3, 0, 0}, {0, 0, 1}));
-            auto h = m.get_histogram()->get_weighted_counts();
-
-            auto b = m.get_body(0).symmetry().explicit_structure();
-            Molecule m2({Body{std::move(b.atoms), std::move(b.waters)}, Body{std::vector{a2}}});
-            set_unity_charge(m2);
-            auto h2 = m2.get_histogram()->get_weighted_counts();
-            CHECK(compare_hist_approx(h, h2));
-        }
-    }
-
-    SECTION("multi-atom body with rotation (explicit_structure comparison)") {
-        // Non-trivial: cm is between atoms so rotation moves copies away from simple translation
-        AtomFF a1({ 1, 0, 0}, form_factor::form_factor_t::C);
-        AtomFF a2({-1, 0, 0}, form_factor::form_factor_t::C);
-        Molecule m({Body{std::vector{a1, a2}}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        SECTION("rotation about y-axis") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({5, 0, 0}, {0, 1, 0}));
-            auto h = m.get_histogram()->get_weighted_counts();
-
-            auto b = m.get_body(0).symmetry().explicit_structure();
-            auto m2 = Molecule({Body{std::move(b.atoms), std::move(b.waters)}});
-            set_unity_charge(m2);
-            auto h2 = m2.get_histogram()->get_weighted_counts();
-            CHECK(compare_hist_approx(h, h2));
-        }
-
-        SECTION("rotation about z-axis") {
-            m.get_body(0).symmetry().add(make_unique_point_sym({0, 3, 0}, {0, 0, 1}));
-            auto h = m.get_histogram()->get_weighted_counts();
-
-            auto b = m.get_body(0).symmetry().explicit_structure();
-            auto m2 = Molecule({Body{std::move(b.atoms), std::move(b.waters)}});
-            set_unity_charge(m2);
-            auto h2 = m2.get_histogram()->get_weighted_counts();
-            CHECK(compare_hist_approx(h, h2));
-        }
-    }
-};
-
-TEST_CASE("SymmetryManager: PointSymmetry") {
-    settings::molecule::implicit_hydrogens = false;
-    settings::molecule::center = false;
-    SECTION("SymmetryManager") {
-        test_point_symmetry(settings::hist::HistogramManagerChoice::HistogramSymmetryManagerMT);
-    }
-    SECTION("PartialSymmetryManager") {
-        test_point_symmetry(settings::hist::HistogramManagerChoice::PartialHistogramSymmetryManagerMT);
-    }
-}
-
-auto test_polyhedral_symmetry = [] (settings::hist::HistogramManagerChoice choice) {
-    static std::random_device rd;
-    static std::mt19937 gen(rd());
-    static std::uniform_real_distribution<> d(-10, 10);
-    static std::uniform_real_distribution<> r(-std::numbers::pi, std::numbers::pi);
-
-    auto make_group = GENERATE(
-        +[] () -> std::unique_ptr<symmetry::PolyhedralSymmetry> {return std::make_unique<symmetry::TetrahedralSymmetry>();},
-        +[] () -> std::unique_ptr<symmetry::PolyhedralSymmetry> {return std::make_unique<symmetry::OctahedralSymmetry>();},
-        +[] () -> std::unique_ptr<symmetry::PolyhedralSymmetry> {return std::make_unique<symmetry::IcosahedralSymmetry>();}
-    );
-    int n_atoms = GENERATE(1, 4);
-
-    // build a body, apply a polyhedral symmetry, and compare the reused-distance histogram
-    // against the ground truth obtained by explicitly materialising every copy
-    for (int i = 0; i < 3; ++i) {
-        std::vector<AtomFF> atoms;
-        for (int j = 0; j < n_atoms; ++j) {
-            atoms.push_back(AtomFF({d(gen), d(gen), d(gen)}, form_factor::form_factor_t::C));
-        }
-
-        Molecule m({Body{atoms}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        auto sym = make_group();
-        sym->translation = {d(gen), d(gen), d(gen)};
-        sym->rotation = {r(gen), r(gen), r(gen)};
-        m.get_body(0).symmetry().add(std::move(sym));
-
-        auto h = m.get_histogram()->get_weighted_counts();
-
-        auto b = m.get_body(0).symmetry().explicit_structure();
-        auto m2 = Molecule({Body{std::move(b.atoms), std::move(b.waters)}});
-        set_unity_charge(m2);
-        auto h2 = m2.get_histogram()->get_weighted_counts();
-
-        CHECK(compare_hist_approx(h, h2));
-    }
-};
-
-TEST_CASE("SymmetryManager: PolyhedralSymmetry") {
-    settings::molecule::implicit_hydrogens = false;
-    settings::molecule::center = false;
-    SECTION("SymmetryManager") {
-        test_polyhedral_symmetry(settings::hist::HistogramManagerChoice::HistogramSymmetryManagerMT);
-    }
-    SECTION("PartialSymmetryManager") {
-        test_polyhedral_symmetry(settings::hist::HistogramManagerChoice::PartialHistogramSymmetryManagerMT);
-    }
-}
-
-auto test_composite_symmetry = [] (settings::hist::HistogramManagerChoice choice) {
-    static std::random_device rd;
-    static std::mt19937 gen(rd());
-    static std::uniform_real_distribution<> d(-8, 8);
-
-    // build inner/outer sub-symmetries; the nesting "inner-outer" replicates the inner unit
-    enum class kind {p2_c3, c2_c3, c3_c2};
-    auto nesting = GENERATE(kind::p2_c3, kind::c2_c3, kind::c3_c2);
-    int n_atoms = GENERATE(1, 3);
-
-    auto make_cyclic = [](double angle, int reps, Vector3<double> offset) {
-        return std::make_unique<symmetry::CyclicSymmetry>(
-            symmetry::CyclicSymmetry::_Relation{offset},
-            symmetry::CyclicSymmetry::_Repeat{{0, 0, 1}, angle},
-            reps
-        );
-    };
-
-    for (int i = 0; i < 3; ++i) {
-        std::vector<AtomFF> atoms;
-        for (int j = 0; j < n_atoms; ++j) {
-            atoms.push_back(AtomFF({d(gen), d(gen), d(gen)}, form_factor::form_factor_t::C));
-        }
-
-        Molecule m({Body{atoms}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        std::unique_ptr<symmetry::ISymmetry> inner, outer;
-        switch (nesting) {
-            case kind::p2_c3:
-                inner = std::make_unique<symmetry::PointSymmetry>(Vector3<double>{4, 1, 0}, Vector3<double>{0, 0, 0});
-                outer = make_cyclic(2*std::numbers::pi/3, 2, {6, 0, 0});
-                break;
-            case kind::c2_c3:
-                inner = make_cyclic(std::numbers::pi, 1, {3, 0, 0});
-                outer = make_cyclic(2*std::numbers::pi/3, 2, {7, 0, 0});
-                break;
-            case kind::c3_c2:
-                inner = make_cyclic(2*std::numbers::pi/3, 2, {3, 0, 0});
-                outer = make_cyclic(std::numbers::pi, 1, {8, 0, 0});
-                break;
-        }
-        m.get_body(0).symmetry().add(std::make_unique<symmetry::CompositeSymmetry>(std::move(inner), std::move(outer)));
-
-        auto h = m.get_histogram()->get_weighted_counts();
-
-        auto b = m.get_body(0).symmetry().explicit_structure();
-        auto m2 = Molecule({Body{std::move(b.atoms), std::move(b.waters)}});
-        set_unity_charge(m2);
-        auto h2 = m2.get_histogram()->get_weighted_counts();
-
-        CHECK(compare_hist_approx(h, h2));
-    }
-};
-
-TEST_CASE("SymmetryManager: CompositeSymmetry") {
-    settings::molecule::implicit_hydrogens = false;
-    settings::molecule::center = false;
-    SECTION("SymmetryManager") {
-        test_composite_symmetry(settings::hist::HistogramManagerChoice::HistogramSymmetryManagerMT);
-    }
-    SECTION("PartialSymmetryManager") {
-        test_composite_symmetry(settings::hist::HistogramManagerChoice::PartialHistogramSymmetryManagerMT);
-    }
-}
-
-auto test_reference_symmetry = [] (settings::hist::HistogramManagerChoice choice) {
-    static std::random_device rd;
-    static std::mt19937 gen(rd());
-    static std::uniform_real_distribution<> d(-8, 8);
-
-    // a cyclic symmetry shared by two bodies, replicating the {body0, body1} pair as a unit
-    auto[angle, reps] = GENERATE(
-        std::make_pair(std::numbers::pi, 1),         // shared c2
-        std::make_pair(2*std::numbers::pi/3, 2)      // shared c3
-    );
-    int n_atoms = GENERATE(1, 3);
-
-    for (int i = 0; i < 3; ++i) {
-        auto random_atoms = [&](int n) {
-            std::vector<AtomFF> atoms;
-            for (int j = 0; j < n; ++j) {atoms.push_back(AtomFF({d(gen), d(gen), d(gen)}, form_factor::form_factor_t::C));}
-            return atoms;
-        };
-        Molecule m({Body{random_atoms(n_atoms)}, Body{random_atoms(n_atoms+1)}});
-        m.set_histogram_manager(choice);
-        set_unity_charge(m);
-
-        symmetry::CyclicSymmetry base(
-            symmetry::CyclicSymmetry::_Relation{{6, 0, 0}},
-            symmetry::CyclicSymmetry::_Repeat{{0, 0, 1}, angle},
-            reps
-        );
-        // body 0 owns the shared symmetry; body 1 holds a view onto it (located by body+slot)
-        m.get_body(0).symmetry().add(std::make_unique<symmetry::ReferenceSymmetry>(base, std::vector<int>{0, 1}, std::vector<int>{0, 0}, &m));
-        m.get_body(1).symmetry().add(std::make_unique<symmetry::ReferenceSymmetryView>(&m, 0, 0));
-
-        auto h = m.get_histogram()->get_weighted_counts();
-
-        // ground truth: materialise every copy of both bodies explicitly
-        auto b0 = m.get_body(0).symmetry().explicit_structure();
-        auto b1 = m.get_body(1).symmetry().explicit_structure();
-        Molecule m2({Body{std::move(b0.atoms), std::move(b0.waters)}, Body{std::move(b1.atoms), std::move(b1.waters)}});
-        set_unity_charge(m2);
-        auto h2 = m2.get_histogram()->get_weighted_counts();
-
-        CHECK(compare_hist_approx(h, h2));
-    }
-};
-
-TEST_CASE("SymmetryManager: ReferenceSymmetry") {
-    settings::molecule::implicit_hydrogens = false;
-    settings::molecule::center = false;
-    SECTION("SymmetryManager") {
-        test_reference_symmetry(settings::hist::HistogramManagerChoice::HistogramSymmetryManagerMT);
-    }
-    SECTION("PartialSymmetryManager") {
-        test_reference_symmetry(settings::hist::HistogramManagerChoice::PartialHistogramSymmetryManagerMT);
     }
 }
